Abstracts a GlobalHotKeyProvider that is an adapter for jkeymaster library
